Update questão 02 b).py

+++ questão 02 b).py
@@ -1,7 +1,7 @@
 m = [[0, 0, 0], [0, 0, 0]   #comecei criando a matriz com os números escolhidos pela pessoa.
 for l in range(0, 2):
     for c in range(0, 3):
-        m[l][c] = int(input(f'Digite um valor para [{l}, {c}]: '))         #usando for in range faz o cog ficar menor, não precisando utilizar append.
+        m[l][c] = float(input(f'Digite um valor para [{l}, {c}]: '))         #usando for in range faz o cog ficar menor, não precisando utilizar append.
 for l in range(0, 2):
     for c in range(0, 3):
         print(f'[{m[l][c]}]', end='')  #saida da matriz em tela.
